+++ src/main/java/com/tagmycode/plugin/gui/form/SettingsForm.java
@@ -3,14 +3,19 @@ package com.tagmycode.plugin.gui.form;
 import com.tagmycode.plugin.BackgroundWorker;
 import com.tagmycode.plugin.Framework;
 import com.tagmycode.plugin.MD5Util;
+import com.tagmycode.plugin.exception.TagMyCodeStorageException;
 import com.tagmycode.plugin.gui.AbstractDialog;
+import com.tagmycode.plugin.gui.ThemeItem;
 import com.tagmycode.sdk.model.User;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 import java.awt.image.BufferedImage;
 import java.net.URL;
+import java.util.ArrayList;
 
 import static com.tagmycode.plugin.gui.GuiUtil.addClickableLink;
 import static com.tagmycode.plugin.gui.GuiUtil.setBold;
@@ -23,6 +28,8 @@ public class SettingsForm extends AbstractDialog {
     private JButton logoutButton;
     private JLabel profilePicture;
     private JButton closeButton;
+    private JComboBox themeComboBox;
+    private JComboBox fontSizeComboBox;
 
     public SettingsForm(final Framework framework, Frame parent) {
         super(framework, parent);
@@ -47,15 +54,98 @@ public class SettingsForm extends AbstractDialog {
     @Override
     protected void initWindow() {
         getDialog().getRootPane().setDefaultButton(null);
-        getDialog().setSize(450, 300);
+        getDialog().setMinimumSize(new Dimension(450, 300));
         getDialog().setResizable(true);
 
+        initGeneralTab();
+        initAccountTab();
+    }
+
+    private void initAccountTab() {
         new BackgroundWorker(new Runnable() {
             @Override
             public void run() {
                 loadProfilePicture(framework.getData().getAccount().getEmail());
             }
         }).execute();
+    }
+
+    private void initGeneralTab() {
+        configureThemesComboBox();
+        configureFontSizeComboBox();
+        ActionListener actionListener = createEditorActionListener();
+        themeComboBox.addActionListener(actionListener);
+        fontSizeComboBox.addActionListener(actionListener);
+    }
+
+    private void configureFontSizeComboBox() {
+        fontSizeComboBox.addItem(8);
+        fontSizeComboBox.addItem(9);
+        fontSizeComboBox.addItem(10);
+        fontSizeComboBox.addItem(12);
+        fontSizeComboBox.addItem(13);
+        fontSizeComboBox.addItem(14);
+        fontSizeComboBox.addItem(16);
+        fontSizeComboBox.addItem(18);
+        fontSizeComboBox.addItem(24);
+        fontSizeComboBox.addItem(36);
+        fontSizeComboBox.addItem(48);
+        fontSizeComboBox.setSelectedItem(framework.getSyntaxSnippetEditorFactory().getFontSize());
+    }
+
+    private ActionListener createEditorActionListener() {
+        return new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent event) {
+                String themeFile = ((ThemeItem) themeComboBox.getSelectedItem()).getKey();
+                try {
+                    int editorFontSize = (int) fontSizeComboBox.getSelectedItem();
+                    framework.getStorageEngine().saveEditorTheme(themeFile);
+                    framework.getStorageEngine().saveEditorFontSize(editorFontSize);
+                    framework.getSyntaxSnippetEditorFactory().applyChanges(themeFile, editorFontSize);
+                } catch (TagMyCodeStorageException e) {
+                    framework.logError(e);
+                }
+            }
+        };
+    }
+
+    private void configureThemesComboBox() {
+        ArrayList<ThemeItem> themes = createThemesArray();
+        ThemeItem loadedTheme = loadTheme(themes);
+
+        for (ThemeItem theme : themes) {
+            themeComboBox.addItem(theme);
+        }
+        if (loadedTheme != null) {
+            themeComboBox.setSelectedItem(loadedTheme);
+        }
+    }
+
+    private ThemeItem loadTheme(ArrayList<ThemeItem> themes) {
+        ThemeItem loadedTheme = null;
+        try {
+            String loadedThemeFile = framework.getStorageEngine().loadEditorTheme();
+            for (ThemeItem theme : themes) {
+                if (theme.getKey().equals(loadedThemeFile)) {
+                    loadedTheme = theme;
+                }
+            }
+        } catch (TagMyCodeStorageException e) {
+            framework.logError(e);
+        }
+        return loadedTheme;
+    }
+
+    private ArrayList<ThemeItem> createThemesArray() {
+        ArrayList<ThemeItem> themes = new ArrayList<>();
+        themes.add(new ThemeItem("default.xml", "Default"));
+        themes.add(new ThemeItem("idea.xml", "IntelliJ IDEA"));
+        themes.add(new ThemeItem("eclipse.xml", "Eclipse"));
+        themes.add(new ThemeItem("vs.xml", "Visual Studio"));
+        themes.add(new ThemeItem("monokai.xml", "Dark Monokai"));
+        themes.add(new ThemeItem("dark.xml", "Dark Obsidian PyCs"));
+        return themes;
     }
 
     @Override